refactor(S5-W): extract src/http.rs + src/util.rs from blocks/helpers.rs
Generic HTTP/response and time/encoding/record/form helpers had nothing
block-specific left after the earlier waves slimmed blocks/helpers.rs, yet
infrastructure (routing, pipeline, ui) reached DOWN into blocks/ for them.

Move them to crate-root modules so the dependency direction reads
infra -> util and blocks -> util:
  - src/http.rs: response sugar. ResponseBuilder + ok_json/ok_empty +
    err_bad_request/unauthorized/forbidden/not_found/conflict/internal/
    internal_no_cause now RE-EXPORT wafer_block::response (the producer half
    landed in phase 1; a local copy of an upstream surface is a shim). Only
    redirect() stays local (solobase-specific convenience over ResponseBuilder).
  - src/util.rs: now_rfc3339/now_millis, stamp_created/updated, path_param,
    json_map/json_as_i64/u64, RecordExt, field_as_string, is_admin,
    block_request/forward_auth_meta, url_path_encode/urlencode,
    parse_form_body/parse_body_value, hex_encode/sha256/sha256_hex re-exports.

blocks/helpers.rs is a TRANSITIONAL re-export shim for this commit only; the
final commit deletes it and points every call site at crate::http / crate::util
directly (no-compat-shims rule).
